requirements.txt
Flask==3.0.0
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
WTForms==3.1.1
Flask-WTF==1.2.1
openai>=1.0.0
